internal/concurrent: store node kind in pointer for HashTrieMap
This is an optimization to save ~8 bytes per node. The core idea is to
use the fact that interior pointers are always valid to the GC,
regardless of whether they refer to validly-typed memory. If we give
node values a high alignment, we can stuff information about them, such
as whether or not they are an entry or indirect node, in their pointers
instead of in the nodes themselves.

Change-Id: I8ef9b8283618fc5fdcb26b77e3555f4de2031d10