internal/concurrent: eliminate parent pointer in HashTrieMap
Currently HashTrieMap tracks parent pointers in all of its indirect
nodes, but these parent pointers are only used for deletion. Deletion
can just as easily track the nodes it encountered on the walk down. If
it doesn't encounter a node as dead after acquiring the lock, we can be
sure all the parent nodes haven't been deleted (and relations are
in-tact) because deletions always happen from leaf nodes up.

Change-Id: I5250522b39bc4aa08531d7d24ab6992fc045af1e