os: document concurrency unsafety of File, Readdir et al
The non-concurrency-safety of the methods of File is
documented explicitly, even though this is unnecessary
according to  https://go.dev/doc/comment#type, because
users may wrongly assume they are merely system calls.

The Readdir method and its variants are also explicitly
documented as unsafe for concurrent use.
(Perhaps this is too much.)

Fixes #66498

Change-Id: I05c88b4e60b44c702062e99ed8f4a32e7945927a