make.bash: tolerate ksh
With this tiny change, we can use Korn shell for a simple build. Specifically,
this has been tested for compiling the latest Go on a clean default install of
OpenBSD with no dependencies except GOROOT_BOOTSTRAP.

What's happening is that
    foo() {
        "$@"
    }
    export R=r
    R=s foo date
    echo $R
prints "r" in bash but "s" in ksh. Adding parens does the right thing for setting
GOROOT under both shells at the small cost of starting a subshell for just the dist build.

This is not intended to imply broader support for anything but bash. Indeed, even
this version of make.bash will complain of an unrecognized -a option to type in
line 167. Also, the script has to be invoked as "ksh make.bash" to avoid a problem
at line 1. Our goal here is to avoid any new maintenance burden for the Go team while
enabling a minimal-dependencies environment for supply chain security researchers.

Updates #3235

Change-Id: I87fc26c12170adf04cea75a30f8b7c06c4fda35b