go/types, types2: allow range over untyped constants representable as integers
In a for-range loop with a numeric range clause, the type of the range
expression x and iteration variable i is determined as in an assignment:

for-range                 corresponding assignment
--------------------------------------------------
var i T                   var i T
for i  = range x {}  <=>  i = x

for i := range x {}  <=>  i := x

for      range x {}  <=>  _ = x // must be integer

This correspondence informs the types of i and x which in the end
must be integer types. If x is an untyped constant, it must fit
into an int.

This change makes for-range work also for all untyped numeric constants x
that can be assigned to an int, such as: for range 1e3 {}.

Fixes #66561.

Change-Id: I682b0e9227e065d6bbd199871c2e1ecff13edc66