internal/maps: add rehash in place
In practice it is rather difficult to achieve >1/3rd of slots as
tombstones due to the delete optimization to mark a deleted slot as
empty if there are other empty slots in the group.

TestTableRehashInPlace implements the most obvious way for this to
occur: with enough inserts, it is likely that at least one group will
become full. If the map is completely cleared at that point then every
slot in the group becomes a tombstone. Repeat this process to create
even more tombstones.

For #54766.

Change-Id: Ib6e979baa5d59ee872cf5ddc80b4fc458df74920