runtime: make profstackdepth a GODEBUG option
This will allow users to decrease the profiling stack depth back to 32
in case they experience any problems with the new default of 128. For
example, this could be used for go programs that expect
StackRecord.Stack() to return no more than 32 frames or that the
returned slice is always backed by the Stack0 array.

Change-Id: Ieaab2513024915a223239278dd97a6e161dde1cf