cmd/compile/internal: support for address-taken vars in stack slot merging
[DO NOT SUBMIT]

WIP, not currently safe (needs more support). This is an experimental
set of changes to flag addr-taken vars reparented into a given
function as a results of inlining, so that we can special-case them
later on in stack slot merging.

The missing piece here is that at the point where we do the
reparenting, we don't have access to the escape analysis disposition
for the var (since if the function being inlined is in another
package, we don't get this info from export data, and if the func is
local, escape analysis hasn't run yet).

Change-Id: I968d8a69962ac1d586ac78b226cc4d7f98a489f6