runtime: don't fetch c-archive/c-shared os.Args on musl/uclinux
This works around the segfault seen when compiling a shared archive and linking it on a musl/uclinux system.

The issue is that the go runtime is doing some magic to extract the program arguments even though the go side is compiled as a library.

By including my changes the glibc behavior remains unchanged, and for musl/uclinux systems we simply do not get any arguments if we are linked as a shared library.

I've attempted to get these changes included in the past and have been manually patching my go releases for alpine.

I believe the issue raised previously was that in theory we could parse the arguments from /proc but honestly I don't have time to mess with that and don't see the need for a library to be aware of argc/argv unless someone here thinks otherwise.

Change-Id: Id8de9d2c21e97ac0ee28328dc69643593aaf54f5
GitHub-Last-Rev: 42ed330c4992289e67282008766f62f741c2c327
GitHub-Pull-Request: golang/go#67254