cmd/compile,cmd/asm: on PPC64, generate compares against constant 0
The R/R/R variant is matched instead of the R/I/R variant when the
second argument is $0 for CMP-like opcodes. This is a matching quirk
whereby R0 and $0 can be used interchangeably, and the R/R/R variant
matches before the R/I/R variant.

The R/I/R variant of CMP can be fused in many cases for better
performance.

Likewise, update a common case to use $0 instead of R0 to take
advantage of this.

Change-Id: If2549ca25a5f7d23001885ad444c70d829b3b066