all: add push linknames to allow legacy pull linknames
CL 585358 adds restrictions to disallow pull-only linknames
(currently off by default). Currently, there are quite some pull-
only linknames in user code in the wild. In order not to break
those, we add push linknames to allow them to be pulled. This CL
includes linknammes found in a large code corpus (thanks Matthew
Dempsky and Michael Pratt!), that are not currently linknamed.

Change-Id: I32f5fc0c7a6abbd7a11359a025cfa2bf458fe767